define some integers in fnalfs computations as int64_t
The extended-range arithmetic method of Prof. Fukushima computes fnalfs
using integers as some auxiliary variables.  These were defined
previously as "int".  With NEON, they need to have the same bit width as
float or double, depending on the precision used to compile CHarm.  This
necessitated to define these integers as "int64_t".

More specifically, with Intel SIMD, we load the integers by explicitly
providing the values of the integers, which are then cast to "int64_t"
when calling the AVX-type intrinsics.  In this case, it is therefore OK
to define these integers as "int".

With NEON, however, we load the integers by referring to a pointer to an
array of integers.  This time, the array has to be therefore of the
"int64_t" type to ensure we load the proper amount of bits for each
integer.  Using "int" with the integers and NEON SIMD intrinsics would
be wrong.

+++ src/leg/leg_func_prepare.c
@@ -1,6 +1,7 @@
 /* Header files */
 /* ------------------------------------------------------------------------- */
 #include <config.h>
+#include <stdint.h>
 #include <math.h>
 #include "../prec.h"
 #include "../simd/simd.h"
@@ -15,7 +16,10 @@
 /* Computes sectorial Legendre functions for internal purposes.  The function
  * is based on the "alfsp" Fortran subroutine (Table 2) due to Fukushima
  * (2012). */
-void CHARM(leg_func_prepare)(const REAL *u, REAL *ps, int *ips, const REAL *dm,
+void CHARM(leg_func_prepare)(const REAL *u,
+                             REAL *ps,
+                             int64_t *ips,
+                             const REAL *dm,
                              unsigned long nmax)
 {
     /* No sectorial Legendre functions for maximum harmonic degree "0" */
@@ -24,7 +28,7 @@ void CHARM(leg_func_prepare)(const REAL *u, REAL *ps, int *ips, const REAL *dm,
 
 
     REAL y[SIMD_SIZE], x[SIMD_SIZE];
-    int ix[SIMD_SIZE];
+    int64_t ix[SIMD_SIZE];
     for (size_t v = 0; v < SIMD_SIZE; v++)
     {
          x[v] = ROOT3 * u[v];
